requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#	production that isn't in development.
-r base.txt


# WSGI Handler
# ------------------------------------------------
gevent==1.0.2

gunicorn==19.4.1

# Static and Media Storage
# ------------------------------------------------
boto==2.38.0
django-storages-redux==1.3








